Make Renderer::setResolution use int BaseType

+++ NAS2D/Renderer/Renderer.h
@@ -127,7 +127,7 @@ public:
 
 	virtual void setViewport(const Rectangle<int>& viewport) = 0;
 	virtual void setOrthoProjection(const Rectangle<float>& orthoBounds) = 0;
-	void setResolution(const Vector<float>& newResolution);
+	void setResolution(const Vector<int>& newResolution);
 
 protected:
 	Renderer(const std::string& appTitle);
